third_party/WebKit/LayoutTests/animations/responsive/animations-responsive-offsetRotate-expected.txt

This is a testharness.js-based test.
FAIL offsetRotate responsive to inherited offsetRotate changes assert_equals: expected "auto 300deg" but got "auto 500deg"
Harness: the test ran to completion.
